ItemList

+++ DX2D/DX_1600/DX_1600/Object/Weapon/Sword.h
@@ -3,7 +3,7 @@
 class Sword : public Weapon
 {
 public:
-	Sword(ItemType type, string name, int price, string itmeDesc, wstring iconSrvfile, int ID, int minAtk, int maxAtk, int atkPerSec, wstring string, Vector2 size);
+	Sword(ItemType type, string name, int price, string itmeDesc, wstring iconSrvfile, int ID, int minAtk, int maxAtk, float atkPerSec, wstring string, Vector2 size);
 	virtual ~Sword();
 	
 	virtual void Update() override;
